# Sprint1_API_Documentation.yaml
openapi: 3.0.0
info:
  title: AgriFlow API
  description: |
    AgriFlow is an agricultural e-commerce platform that connects farmers with customers, 
    enabling product listing, ordering, cart management, booking, reviews, and administrative 
    features including demand forecasting, recommendations, credit scoring, and offer management.
    
    ## User Story Mapping
    This API implements all user stories from Sprint 1:
    - US1: User Registration & Login
    - US2: Profile Management
    - US3: Browse Products
    - US4: Search & Filter
    - US5: Add Products (Admin)
    - US6: Update Products (Admin)
    - US7: Delete Products (Admin)
    - US8: Cart Management
    - US9: Place Order
    - US10: Order History
    - US11: Product Reviews (after purchase)
    - US12: View Reviews
    - US13: Address Management
    - US14: Admin Dashboard
    - US15: Expiry Alerts
    - US16: Top Customers
    - US17: Monthly Reports
    - US18: AI Demand Forecasting (GenAI Integration)
    - US19: Credit Score Management
    - US20: Payment Tracking
    - US21: Offers Management
    - US22: Review Analytics
    - US23: Offer Notifications
    - US24: Smart Recommendations
    
    ## Integrated GenAI APIs
    - **Demand Forecasting API**: Uses linear regression and trend analysis to predict future product demand
  version: 1.0.0
  contact:
    name: AgriFlow Developer Team
servers:
  - url: http://localhost:5000/api
    description: Development server
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account and automatically creates a shopping cart for the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  description: User's full name
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  description: User's password
                mobile_no:
                  type: string
                  description: User's mobile number
                role:
                  type: string
                  enum: [customer, admin]
                  default: customer
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: integer
        400:
          description: Email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates user and returns user details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: integer
                  role:
                    type: string
                  name:
                    type: string
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Returns all products. Customers see only non-expired products, admins see all products.
      parameters:
        - name: Role
          in: header
          schema:
            type: string
            enum: [customer, admin]
          description: User role to determine product visibility
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
    post:
      tags:
        - Products
      summary: Create a new product
      description: Creates a new product (admin only)
      parameters:
        - name: Role
          in: header
          required: true
          schema:
            type: string
            enum: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - quantity
                - price
              properties:
                name:
                  type: string
                description:
                  type: string
                quantity:
                  type: integer
                price:
                  type: number
                expiry_date:
                  type: string
                  format: date
      responses:
        201:
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  product_id:
                    type: integer
        403:
          description: Unauthorized
  /products/{product_id}:
    get:
      tags:
        - Products
      summary: Get a specific product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/Product'
        404:
          description: Product not found
    put:
      tags:
        - Products
      summary: Update a product
      description: Updates product details (admin only)
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
        - name: Role
          in: header
          required: true
          schema:
            type: string
            enum: [admin]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                quantity:
                  type: integer
                price:
                  type: number
                expiry_date:
                  type: string
      responses:
        200:
          description: Product updated
        403:
          description: Unauthorized
    delete:
      tags:
        - Products
      summary: Delete a product
      description: Deletes a product and all related data (admin only)
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
        - name: Role
          in: header
          required: true
          schema:
            type: string
            enum: [admin]
      responses:
        200:
          description: Product deleted
        403:
          description: Unauthorized
  /cart:
    get:
      tags:
        - Cart
      summary: Get user's cart
      parameters:
        - name: User-ID
          in: header
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Cart items and total
          content:
            application/json:
              schema:
                type: object
                properties:
                  cart_items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CartItem'
                  total_price:
                    type: number
        400:
          description: User ID required
        404:
          description: Cart not found
  /cart/add:
    post:
      tags:
        - Cart
      summary: Add item to cart
      parameters:
        - name: User-ID
          in: header
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - product_id
                - quantity
              properties:
                product_id:
                  type: integer
                quantity:
                  type: integer
      responses:
        200:
          description: Item added to cart
        400:
          description: Insufficient stock or User ID required
  /cart/update/{cart_item_id}:
    put:
      tags:
        - Cart
      summary: Update cart item quantity
      parameters:
        - name: cart_item_id
          in: path
          required: true
          schema:
            type: integer
        - name: User-ID
          in: header
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
      responses:
        200:
          description: Cart item updated
        400:
          description: Insufficient stock or User ID required
        403:
          description: Unauthorized
  /cart/remove/{cart_item_id}:
    delete:
      tags:
        - Cart
      summary: Remove item from cart
      parameters:
        - name: cart_item_id
          in: path
          required: true
          schema:
            type: integer
        - name: User-ID
          in: header
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Item removed from cart
        400:
          description: User ID required
        403:
          description: Unauthorized
...
  /bookings/history:
    get:
      tags:
        - Bookings
      summary: Get booking history
      description: Returns user's order history (or all orders for admin)
      parameters:
        - name: User-ID
          in: header
          required: true
          schema:
            type: integer
        - name: Role
          in: header
          schema:
            type: string
            enum: [customer, admin]
      responses:
        200:
          description: List of bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
        400:
          description: User ID required
...
  /reviews/product/{product_id}:
    get:
      tags:
        - Reviews
      summary: Get product reviews
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: List of reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
...